amd: Dynamo maturin build PASS, driver mismatch identified for disagg
Dynamo native extension (maturin):
- Built successfully in rocm/sgl-dev container (Rust 1.93.1 + maturin 1.12.4)
- dynamo._core version 1.0.0 installed
- Fix: set default = ["cuda"] in dynamo-memory and kvbm-physical Cargo.toml
- Fix: BINDGEN_EXTRA_CLANG_ARGS for stdbool.h
- Fix: cudarc device::get() takes i32 not usize in gpu/cuda.rs

SGLang MoRI disagg:
- Container pulled on both nodes (rocm/sgl-dev:sglang-0.5.9-rocm720-mi35x-mori-0227-2)
- BLOCKED by ionic RDMA driver ABI mismatch (host ABI 1 vs container ABI 4)
- Same issue affects RIXL/UCX inter-node — all need matching AINIC driver
- Resolution: update AINIC driver on chi2899/chi2900

Updated test results and disagg serving guide with driver findings.

Made-with: Cursor

+++ lib/memory/src/gpu/cuda.rs
@@ -50,7 +50,7 @@ impl super::GpuDevice for CudaBackend {
     }
 
     fn create_context(device_id: i32) -> GpuResult<ContextHandle> {
-        let dev = result::device::get(device_id as usize).map_err(cuda_err)?;
+        let dev = result::device::get(device_id).map_err(cuda_err)?;
         let mut ctx: sys::CUcontext = std::ptr::null_mut();
         check_cu(unsafe { sys::cuDevicePrimaryCtxRetain(&mut ctx, dev) })?;
         Ok(ContextHandle(ctx as *mut c_void))
@@ -190,7 +190,7 @@ impl super::GpuDevice for CudaBackend {
     }
 
     fn device_name(device_id: i32) -> GpuResult<String> {
-        let dev = result::device::get(device_id as usize).map_err(cuda_err)?;
+        let dev = result::device::get(device_id).map_err(cuda_err)?;
         let mut name_buf = [0i8; 256];
         check_cu(unsafe {
             sys::cuDeviceGetName(name_buf.as_mut_ptr(), name_buf.len() as i32, dev)
@@ -200,7 +200,7 @@ impl super::GpuDevice for CudaBackend {
     }
 
     fn total_memory(device_id: i32) -> GpuResult<usize> {
-        let dev = result::device::get(device_id as usize).map_err(cuda_err)?;
+        let dev = result::device::get(device_id).map_err(cuda_err)?;
         let mut bytes: usize = 0;
         check_cu(unsafe { sys::cuDeviceTotalMem_v2(&mut bytes, dev) })?;
         Ok(bytes)
